+++ Template/Tag/MicrosoftConsentModeTag.php
@@ -12,7 +12,7 @@ use Piwik\Settings\FieldConfig;
 use Piwik\Plugins\TagManager\Template\Tag\BaseTag;
 use Piwik\Validators\NotEmpty;
 
-class GoogleConsentModeTag extends BaseTag
+class MicrosoftConsentModeTag extends BaseTag
 {
     public function getName()
     {
@@ -51,25 +51,19 @@ class GoogleConsentModeTag extends BaseTag
         // return parent::getIcon();
         //
         // The image should have ideally a resolution of about 64x64 pixels.
-        return 'plugins/TagManagerExtended/images/icons/tag/google-consent-mode.svg';
+        return 'plugins/TagManagerExtended/images/icons/tag/microsoft-advertising.png';
     }
 
     public function getParameters()
     {
         $default = [
             ['consent_type' => 'ad_storage', 'consent_state' => 'granted'],
-            ['consent_type' => 'ad_user_data', 'consent_state' => 'granted'],
-            ['consent_type' => 'ad_personalization', 'consent_state' => 'granted'],
-            ['consent_type' => 'analytics_storage', 'consent_state' => 'granted'],
-            ['consent_type' => 'functionality_storage', 'consent_state' => 'granted'],
-            ['consent_type' => 'personalization_storage', 'consent_state' => 'granted'],
-            ['consent_type' => 'security_storage', 'consent_state' => 'granted'],
         ];
 
         return array(
             $this->makeSetting('consentAction', 'update', FieldConfig::TYPE_ARRAY, function (FieldConfig $field) {
-                $field->title = Piwik::translate('TagManagerExtended_GoogleConsentModeTagConsentActionTitle');
-                $field->description = Piwik::translate('TagManagerExtended_GoogleConsentModeTagConsentActionDescription');
+                $field->title = Piwik::translate('TagManagerExtended_MicrosoftConsentModeTagConsentActionTitle');
+                $field->description = Piwik::translate('TagManagerExtended_MicrosoftConsentModeTagConsentActionDescription');
                 $field->uiControl = FieldConfig::UI_CONTROL_SINGLE_SELECT;
                 $field->availableValues = array(
                     'default' => 'default',
@@ -80,24 +74,17 @@ class GoogleConsentModeTag extends BaseTag
 
             $this->makeSetting('consentTypes', $default, FieldConfig::TYPE_ARRAY, function (FieldConfig $field) {
                 $field->uiControl = FieldConfig::UI_CONTROL_MULTI_TUPLE;
-                $field->title = Piwik::translate('TagManagerExtended_GoogleConsentModeTagConsentTypesTitle');
-                $field->description = Piwik::translate('TagManagerExtended_GoogleConsentModeTagConsentTypesDescription');
+                $field->title = Piwik::translate('TagManagerExtended_MicrosoftConsentModeTagConsentTypesTitle');
+                $field->description = Piwik::translate('TagManagerExtended_MicrosoftConsentModeTagConsentTypesDescription');
 
-                $field1 = new FieldConfig\MultiPair(Piwik::translate('TagManagerExtended_GoogleConsentModeTagConsentTypeTitle'), 'consent_type', FieldConfig::UI_CONTROL_TEXT);
+                $field1 = new FieldConfig\MultiPair(Piwik::translate('TagManagerExtended_MicrosoftConsentModeTagConsentTypeTitle'), 'consent_type', FieldConfig::UI_CONTROL_TEXT);
                 $field1->customFieldComponent = self::FIELD_VARIABLE_COMPONENT;
 
-                $field2 = new FieldConfig\MultiPair(Piwik::translate('TagManagerExtended_GoogleConsentModeTagConsentStateTitle'), 'consent_state', FieldConfig::UI_CONTROL_TEXT);
+                $field2 = new FieldConfig\MultiPair(Piwik::translate('TagManagerExtended_MicrosoftConsentModeTagConsentStateTitle'), 'consent_state', FieldConfig::UI_CONTROL_TEXT);
                 $field2->customFieldComponent = self::FIELD_VARIABLE_COMPONENT;
 
                 $field->uiControlAttributes['field1'] = $field1->toArray();
                 $field->uiControlAttributes['field2'] = $field2->toArray();
-            }),
-
-            $this->makeSetting('waitForUpdate', 500, FieldConfig::TYPE_STRING, function (FieldConfig $field) {
-                $field->title = Piwik::translate('TagManagerExtended_GoogleConsentModeTagWaitForUpdateTitle');
-                $field->description = Piwik::translate('TagManagerExtended_GoogleConsentModeTagWaitForUpdateDescription');
-                $field->customFieldComponent = self::FIELD_VARIABLE_COMPONENT;
-                $field->validators[] = new NotEmpty();
             }),
         );
     }